add working game module
